+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # Changelog
+
 All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
@@ -6,25 +7,44 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+### Added
+
+- netcoreapp3.1 support
+
+### Changed
+
+- Update IKVM to 8.2
+- Downgrade FeatureIDE library to version 3.3.0
+- Target net461 (required by IKVM)
+- Require nuget v4.3
+- Release with github actions
 
 ## [3.5.5] - 2020-12-23
+
 ### Changed
+
 - Updated nuspec file.
 - Updated FeatureIDE library to version 3.5.5
 
 ## [3.5.2] - 2018-11-28
+
 ### Changed
+
 - Updated FeatureIDE library to version 3.5.2
 
 ## [3.3.0] - 2017-08-21
+
 ### Changed
+
 - Updated FeatureIDE library to version 3.3.0
 
 ## 3.2.0 - 2017-06-21
+
 ### Added
+
 - Initial release based on FeatureIDE library 3.2.0
 
-[Unreleased]: https://github.com/VisualOn/FeatureIDE.net/compare/v3.5.5...HEAD
-[3.5.5]: https://github.com/VisualOn/FeatureIDE.net/compare/v3.5.2...v3.5.5
-[3.5.2]: https://github.com/VisualOn/FeatureIDE.net/compare/v3.3.0...v3.5.2
-[3.3.0]: https://github.com/VisualOn/FeatureIDE.net/compare/v3.2.0...v3.3.0
+[unreleased]: https://github.com/visualon/featureide.net/compare/v3.5.5...HEAD
+[3.5.5]: https://github.com/visualon/featureide.net/compare/v3.5.2...v3.5.5
+[3.5.2]: https://github.com/visualon/featureide.net/compare/v3.3.0...v3.5.2
+[3.3.0]: https://github.com/visualon/featureide.net/compare/v3.2.0...v3.3.0